+++ frontend/src/App.tsx
@@ -3,7 +3,8 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { supabase } from "@/config/db.config.ts";
 import Login from "@/pages/Login";
 import Home from "@/pages/Home";
-import Map from "./pages/map/Map";
+import Map from "@/pages/Map/Map";
+import {SelectedPage} from "@/shared/SelectedPage.ts";
 
 
 type User = {
@@ -13,10 +14,29 @@ type User = {
 };
 
 function App() {
+  const [selectedPage, setSelectedPage] = useState<SelectedPage>(SelectedPage.Home);
   const [users, setUsers] = useState<User[]>([]);
+  const [isTopOfPage, setIsTopOfPage] = useState<boolean>(true);
 
   useEffect(() => {
     getUsers();
+  }, []);
+
+  useEffect(() => {
+
+    const handleScroll = () => {
+      if (window.scrollY === 0) {
+        setIsTopOfPage(true);
+      }
+
+      if (window.scrollY !== 0) {
+        setIsTopOfPage(false);
+      }
+    };
+
+    window.addEventListener('scroll', handleScroll);
+
+    return () => window.removeEventListener('scroll', handleScroll);
   }, []);
 
   async function getUsers() {
@@ -28,10 +48,10 @@ function App() {
       <div className='app bg-primary-2'>
         <Router>
           <Routes>
-            <Route path="/login" element={<Login/>}/>
-            <Route path="/home" element={<Home/>}/>
-            <Route path="/map" element={<Map/>}/>
-            <Route path="/" element={<Home/>}/>
+            <Route path="/login" element={<Login selectedPage={selectedPage} setSelectedPage={setSelectedPage}/>}/>
+            <Route path="/home" element={<Home selectedPage={selectedPage} setSelectedPage={setSelectedPage} isTopOfPage={isTopOfPage}/>} />
+            <Route path="/map" element={<Map selectedPage={selectedPage} setSelectedPage={setSelectedPage} />}/>
+            <Route path="/" element={<Home selectedPage={selectedPage} setSelectedPage={setSelectedPage} isTopOfPage={isTopOfPage}/>}/>
           </Routes>
         </Router>
       </div>
