+++ frontend/src/components/Navbar/Link.tsx
@@ -14,7 +14,7 @@ function Link({ page, selectedPage, setSelectedPage, isAboveMediumScreen }: Prop
         <a
             href={`${lowerCasePage}`}
             onClick={() => setSelectedPage(lowerCasePage)}
-            className={`${(selectedPage === lowerCasePage && isAboveMediumScreen) ? 'text-primary-4' : 'text-black'}
+            className={`${selectedPage === lowerCasePage && isAboveMediumScreen ? 'text-primary-4' : 'text-black'}
                         ${!isAboveMediumScreen && lowerCasePage === SelectedPage.SignIn ? 'text-primary-4' : 'text-black'} font-karla font-medium text-k-16` }
         >
             { page }
